* DefObject.h: Modify setBit by casting the bit as unsigned.
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@2469 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/DefObject.h
@@ -91,7 +91,8 @@ extern id lispInKeyword (id index);
 //
 
 #define getBit(word, bit) ((word) & bit)
-#define setBit(word, bit, value) ((value) ? ((word) |= bit) : ((word) &= ~bit))
+#define setBit(word, bit, value) ((value) ? ((word) |= (unsigned)bit) : 
+				  ((word) &= ~ (unsigned)bit))
 
 #define setField(word, shift, value) (word) |= ((value) << shift)
